+++ src/app/layout.tsx
@@ -1,0 +1,10 @@
+import React, { ReactNode } from 'react'
+
+import _Layout from '../layout'
+
+type Props = {
+  children: ReactNode
+}
+
+const Layout: React.FC<Props> = ({ children }) => <_Layout>{children}</_Layout>
+export default Layout
