--- src/pages/_document.tsx
@@ -1,27 +1,0 @@
-import Document, { DocumentContext, Html, Head, Main, NextScript } from 'next/document'
-
-export default class CustomDocument extends Document {
-  static async getInitialProps(context: DocumentContext) {
-    const initialProps = await Document.getInitialProps(context)
-    return { ...initialProps }
-  }
-  render() {
-    return (
-      <Html lang='ja'>
-        <Head>
-          {/* load external font */}
-          <link rel='preconnect' href='https://fonts.googleapis.com' />
-          <link rel='preconnect' href='https://fonts.gstatic.com' crossOrigin='' />
-          <link
-            href='https://fonts.googleapis.com/css2?family=Signika+Negative:wght@400;700&family=Work+Sans:wght@400;600&display=swap'
-            rel='stylesheet'
-          />
-        </Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    )
-  }
-}
